Update bugsinpy-compile
Not sure if this is a personal problem, but if I run `python venv` on my mac it tries to use Python2 instead of Python3, which causes the compilation to fail with "No module named venv". Changing to python3 fixed this for me. It still fails with "dos2unix: command not found" which I fixed by separately installing dos2unix. This should be listed as a requirement somewhere though or handled in the script.